LayoutTests/imported/w3c/web-platform-tests/svg/animations/animate-elem-18-t-drt-expected.txt

calcMode="discrete"
keyTimes="0;.2;.6;.8"
Time (s):
0
2
6
8
$Revision: 1.1 $

FAIL A copy of the corresponding W3C-SVG-1.1 test, which dumps the animation at certain times assert_approx_equals: expected 30 +/- 1 but got 180
